[pknoth-master] Splitting in two files
